розібрався ззакриванням модалок
