+++ src/index.js
@@ -1,3 +1,8 @@
 import './js/menu';
-import './js/modal';
+import './js/modal-header';
+import './js/modal-mobile-menu';
+import './js/modal-products';
+import './js/modal-about';
+import './js/modal-our-locations';
+import './js/modal-franchise';
 import './js/slider';
