+++ src/js/modal-mobile-menu.js
@@ -1,0 +1,14 @@
+(() => {
+    const refs = {
+        openModalBtn: document.querySelector("[data-modal-open-mobile]"),
+        closeModalBtn: document.querySelector("[data-modal-close-mobile]"),
+        modal: document.querySelector("[data-modal-mobile]"),
+    };
+
+    refs.openModalBtn.addEventListener("click", toggleModal);
+    refs.closeModalBtn.addEventListener("click", toggleModal);
+
+    function toggleModal() {
+        refs.modal.classList.toggle("is-hidden");
+    }
+    })();
