+++ src/js/modal-products.js
@@ -1,35 +1,5 @@
 (() => {
     const refs = {
-      openModalBtn: document.querySelector("[data-modal-open]"),
-      closeModalBtn: document.querySelector("[data-modal-close]"),
-      modal: document.querySelector("[data-modal]"),
-    };
-  
-    refs.openModalBtn.addEventListener("click", toggleModal);
-    refs.closeModalBtn.addEventListener("click", toggleModal);
-  
-    function toggleModal() {
-      refs.modal.classList.toggle("is-hidden");
-    }
-  })();
-
-  (() => {
-    const refs = {
-      openModalBtn: document.querySelector("[data-modal-open-mobile-menu]"),
-      closeModalBtn: document.querySelector("[data-modal-close-mobile-menu]"),
-      modal: document.querySelector("[data-modal-mobile-menu]"),
-    };
-  
-    refs.openModalBtn.addEventListener("click", toggleModal);
-    refs.closeModalBtn.addEventListener("click", toggleModal);
-  
-    function toggleModal() {
-      refs.modal.classList.toggle("is-hidden");
-    }
-  })();
-
-  (() => {
-    const refs = {
       openModalBtn: document.querySelector("[data-modal-open-products-1]"),
       closeModalBtn: document.querySelector("[data-modal-close-products-1]"),
       modal: document.querySelector("[data-modal-products-1]"),
@@ -63,21 +33,6 @@
       openModalBtn: document.querySelector("[data-modal-open-products-3]"),
       closeModalBtn: document.querySelector("[data-modal-close-products-3]"),
       modal: document.querySelector("[data-modal-products-3]"),
-    };
-  
-    refs.openModalBtn.addEventListener("click", toggleModal);
-    refs.closeModalBtn.addEventListener("click", toggleModal);
-  
-    function toggleModal() {
-      refs.modal.classList.toggle("is-hidden");
-    }
-  })();
-
-  (() => {
-    const refs = {
-      openModalBtn: document.querySelector("[data-modal-open-about]"),
-      closeModalBtn: document.querySelector("[data-modal-close-about]"),
-      modal: document.querySelector("[data-modal-about]"),
     };
   
     refs.openModalBtn.addEventListener("click", toggleModal);
