+++ src/partials/header.html
@@ -27,7 +27,7 @@
 					</svg>
 			</button>
 
-			<button type="button" data-modal-open class="button header__button header__button-top">Buy now
+			<button type="button" class="button header__button header__button-top" data-modal-open>Buy now
 				<svg class="header__icon" width="12" height="12">
 						<use class="" href="./images/icons.svg#icon-Vector-1"></use>
 				</svg>
@@ -36,7 +36,7 @@
 			<!-- * Мобильное меню -->
 
 			<div class="mobile-menu is-hidden" data-menu>
-					<button type="button" class="menu__button menu__button-close" data-menu-close>
+					<button type="button" class="menu__button menu__button-close" data-menu-close data-modal-close-mobile>
 							<svg class="" width="32" height="32" aria-label="Вызов мобильного меню">
 									<use class="icon-close" href="./images/icons.svg#menu-shut"></use>
 							</svg>
@@ -62,7 +62,7 @@
 							</ul>
 					</nav>
 
-					<button type="button" data-modal-open-mobile-menu class="button header__button">Buy now
+					<button type="button" class="button header__button" data-modal-open-mobile>Buy now
 							<svg class="header__icon" width="12" height="12">
 									<use class="" href="./images/icons.svg#icon-Vector-1"></use>
 							</svg>
