+++ src/partials/modal.html
@@ -1,8 +1,8 @@
 <!-- ==================================================================HEADER-MODAL================================================ -->
     
-    <div class="backdrop is-hidden" data-modal data-modal-mobile-menu>
+    <div class="backdrop is-hidden" data-modal data-modal-mobile>
         <div class="modal-header">
-            <button type="button" class="close-modal" data-modal-close>
+            <button type="button" class="close-modal"data-modal-close data-modal-close-mobile>
                 <svg class="modal__close-icon" width="20" height="20">
                     <use href="./images/icons.svg#menu-close"></use>
                 </svg>
