+++ src/sass/_header.scss
@@ -148,9 +148,9 @@ color: var(--main-light-color);
 	margin-left: 6px;
 }
 
-// .header {
-// 	// position: relative;
-// }
+.header {
+	// position: relative;
+}
 
 .header__container {
 	display: flex;
