+++ src/sass/_modal-header.scss
@@ -1,0 +1,283 @@
+// ==============================================================HEADER-MODAL==============================================================
+
+.backdrop {
+    position: fixed;
+    z-index: 100;
+    top: 0px;
+    left: 0px ;
+    width: 100%;
+    height: 100%;
+    background: rgba(0, 0, 0, 0.2);
+    transition: opacity 250ms cubic-bezier(0.4, 0, 0.2, 1),
+    visibility 250ms cubic-bezier(0.4, 0, 0.2, 1);
+    
+    &.is-hidden {
+        opacity: 0;
+        visibility: hidden;
+        pointer-events: none;
+    }
+}
+
+.modal-header {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    background: #FFFFFF;
+    box-shadow: 0px 1px 3px rgba(0, 0, 0, 0.12), 0px 1px 1px rgba(0, 0, 0, 0.14), 0px 2px 1px rgba(0, 0, 0, 0.2);
+    border-radius: 10px;
+    padding: 28px 28px;
+    max-height: 98%;
+    overflow-y: auto;
+}
+
+.close-modal {
+    position: absolute;
+    top: 8px;
+    right: 8px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 20px;
+    height: 20px;
+    background-color: transparent;
+    border: none;
+    padding: 0px;
+    cursor: pointer;
+}
+
+.modal-header__title {
+    text-align: center;
+    font-family: 'Seymour One';
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 1.5;
+}
+.modal-header-form {
+    display: flex;
+    flex-direction: column;
+    margin-top: 84px;
+}
+
+.modal-assortment {
+    display: flex;
+}
+
+
+.checkbox-ice-cream:checked + .type-ice-cream {
+    background-color: #96AF8C;
+    transition: background-color 250ms cubic-bezier(0.4, 0, 0.2, 1);
+}
+
+.checkbox-ice-coffee:checked + .type-ice-coffee {
+    background-color: #EFD478;
+    transition: background-color 250ms cubic-bezier(0.4, 0, 0.2, 1);
+}
+
+.checkbox-milkshake:checked + .type-milkshake {
+    background-color: #D6936D;
+    transition: background-color 250ms cubic-bezier(0.4, 0, 0.2, 1);
+}
+
+.type-assortment {
+    display: flex;
+    justify-content: flex-end;
+    flex-direction: column;
+    text-align: center;
+    position: relative;
+    width: 132px;
+    height: 135px;
+    font-weight: 700;
+    font-size: 16px;
+    line-height: 1.25;
+    letter-spacing: 0.04em;
+    text-transform: uppercase;
+    color: var(--main-text-color);
+    padding-bottom: 20px;
+    border-radius: 16px;
+    border: 2px solid;
+    cursor: pointer;
+} 
+
+.modal-asortment-pic {
+    position: absolute;
+    top: -50%;
+    left: -2px;
+    margin: 0 auto;
+    width: 132px;
+    height: 138px;
+    border-radius: 50%;
+}
+
+.type-assortment__image {
+    display: block;
+    position: absolute;
+    top: -3px;
+}
+
+
+.modal-asortment-ice-cream {
+    background: linear-gradient(0deg, rgba(156, 190, 143, 0.7), rgba(156, 190, 143, 0.7)), #D9D9D9;
+    box-shadow: 0px 5.26027px 19.726px rgba(10, 136, 23, 0.25);
+}
+
+.modal-asortment-ice-coffee {
+    background: linear-gradient(0deg, rgba(248, 210, 78, 0.7), rgba(248, 210, 78, 0.7)), #D9D9D9;
+    box-shadow: 0px 5.26027px 19.726px rgba(153, 114, 23, 0.25);
+}
+
+.modal-asortment-milkshake {
+    background: linear-gradient(0deg, rgba(213, 117, 63, 0.7), rgba(213, 117, 63, 0.7)), #D9D9D9;
+    box-shadow: 0px 5.26027px 19.726px rgba(0, 0, 0, 0.25);
+}
+
+.type-ice-cream {
+    border-color: #96AF8C;
+}
+
+.type-ice-coffee {
+    border-color: #EFD478;
+}
+
+.type-milkshake {
+    border-color: #D6936D;
+}
+
+.modal-order-form {
+    text-align: center;
+}
+
+.modal-header-form__input {
+    background-color: var(--input-bg-color);
+    border: 1px solid var(--main-bg-color);
+    color: var(--main-text-color);
+    border-radius: 10px;
+    padding-left: 16px;
+    line-height: 140%;
+    width: 100%;
+    height: 32px;
+    transition: border-color 250ms cubic-bezier(0.4, 0, 0.2, 1);
+
+    &::placeholder {
+        color: var(--main-text-color);
+    }
+    &:focus {
+        outline: none;
+        border-color: red;
+    }
+}
+
+.input-name {
+    margin-top: 24px;
+    }
+
+.input-phone {
+margin-top: 12px;
+}
+
+.modal-header-form__message {
+    background-color: var(--input-bg-color);
+    border: 1px solid var(--main-bg-color);
+    color: var(--main-text-color);
+    border-radius: 10px;;
+    padding-left: 16px;
+    padding-top: 11px;
+    line-height: 140%;
+    width: 100%;
+    height: 64px;
+    resize: none;
+    margin-top: 12px;
+    
+    &::placeholder {
+        color: var(--main-text-color);
+    }
+    &:focus {
+        outline: none;
+    }
+} 
+
+.modal-header-form__submit {
+    font-weight: 600;
+    font-size: 14px;
+    line-height: 17px;
+    width: 114px;
+    height: 42px;
+    border-radius: 10px;
+    color: var(--main-text-color);
+    background-color: var(--accent-color);
+    border: none;
+    padding: 0;
+    margin: 24 auto 0 auto;
+}
+
+
+@media screen and (max-width: 479px){
+    .modal-header {
+        width: 96%;
+    }
+}
+
+@media screen and (min-width: 480px) {
+    .modal-header {
+        width: 450px;
+    }
+}
+
+@media screen and (max-width: 767px) {
+    .modal-assortment {
+        flex-direction: column;
+        align-items:center;
+    }
+    .type-milkshake, .type-ice-coffee {
+        margin-top: 84px;
+    }
+}
+
+@media screen and (min-width: 768px) {
+    .close-modal {
+        top: 20px;
+        right: 20px;
+    }
+    
+    .modal-header {
+        width: 568px;
+        height: 644px;
+        padding: 32px 54px;
+    }
+    
+    .modal-header-form {
+        margin: 0 auto;
+        margin-top: 103px;
+    }
+
+    .modal-assortment {
+        width: 100%; 
+        justify-content: space-between;
+    }
+
+    .modal-header-form__input {
+        width: 368px;
+        height: 42px;
+    }
+
+    .input-name {
+        margin-top: 40px;
+        }
+    
+
+    .input-phone {
+        margin-top: 16px;
+        }
+
+    .modal-header-form__message {
+        width: 368px;
+        height: 84px;
+        margin-top: 16px;
+    }
+
+    .modal-header-form__submit {
+        width: 121px;
+        height: 42px;
+        margin: 32 auto 0 auto;
+    }
+}
