+++ src/sass/index.scss
@@ -10,5 +10,7 @@
 @import './gallery';
 @import './customers';
 @import './contacts';
-@import './modal';
+@import './modal-header';
+@import './modal-products';
+@import './modal-about';
 @import './footer';
